requirements.txt
playwright~=1.51.0
requests~=2.32.3
python-dotenv~=0.21.0
python-telegram-bot~=20.7
fastapi~=0.115.7
uvicorn~=0.34.0
flask~=3.0.3